Reorganise code for disjoint set
and add build entry in CMakeLists
