+++ C++/test/data_structure/disjoint_set.cpp
@@ -1,7 +1,5 @@
-#define DISJOINT_SET_TEST
-
 #include "third_party/catch.hpp"
-#include "data_structures/disjoint_set.cpp"
+#include "data_structure/disjoint_set.hpp"
 
 TEST_CASE("Create a large set containing 10^8 elements", "[disjoint-set]") {
     /*
@@ -93,5 +91,3 @@ TEST_CASE("Connect elements and verify their representatives", "[disjoint-set]")
         REQUIRE(ds.find(element) == 11);
     }
 }
-
-#undef DISJOINT_SET_TEST
